Refactor for multi sensor support
