+++ omotion/Interface.py
@@ -3,11 +3,15 @@ import logging
 
 from omotion.Console import MOTIONConsole
 from omotion.Sensor import MOTIONSensor
-from omotion.core import MOTIONUart, MOTIONSignal, MotionComposite
+from omotion.MotionUart import MOTIONUart
+from omotion.MotionComposite import MotionComposite
+
 from omotion.config import CONSOLE_MODULE_PID, SENSOR_MODULE_PID
 from omotion.signal_wrapper import SignalWrapper, PYQT_AVAILABLE
 
-logger = logging.getLogger(__name__)
+logger = logging.getLogger("Interface")
+logger.setLevel(logging.INFO)
+
 
 class MOTIONInterface(SignalWrapper):
     
@@ -36,7 +40,7 @@ class MOTIONInterface(SignalWrapper):
 
         # Create a MOTIONSensor Device instance as part of the interface
         logger.debug("Initializing Sensor Module of MOTIONInterface with VID: %s, PID: %s, baudrate: %s, timeout: %s", vid, sensor_pid, baudrate, timeout)
-        self._sensor_uart = MotionComposite(vid=vid, pid=sensor_pid, baudrate=baudrate, timeout=timeout, desc="sensor", demo_mode=False, async_mode=run_async)
+        self._sensor_uart = MotionComposite(vid=vid, pid=sensor_pid, timeout=timeout, desc="sensor", demo_mode=False, async_mode=run_async)
         self.sensor_module = MOTIONSensor(uart=self._sensor_uart)
 
         # Connect signals to internal handlers
